hamburger
Added a working hamburger menu to the sticky nav.

Still need to add animations to the list items inside of it so they don't pop in before the opening animation completes. That will be next.

I also changed some variable names to be more descriptive and accurate to what they are.